Recover the retail GT4 image interface
Accept full integer screen coordinates at the public boundary and narrow them once inside SetupImageToPolyGT4. This preserves the demo-era packet behavior while matching the retail callers' ABI.

+++ src/main.exe/images.h
@@ -110,6 +110,6 @@ extern GsIMAGE *GetImage(ImageArchiveId id);
 extern void SetupImageToPolyFT4(GsIMAGE *image, POLY_FT4 *ply,
                                 short x, short y);
 extern void SetupImageToPolyGT4(GsIMAGE *image, POLY_GT4 *ply,
-                                short x, short y);
+                                int x, int y);
 
 #endif
